+++ package.json
@@ -17,9 +17,9 @@
   "scripts": {
     "test": "bash -c 'node_modules/istanbul/lib/cli.js cover node_modules/mocha/bin/_mocha' && npm run browser-test",
     "browser-test": "bash -c 'node_modules/.bin/browserify test/di.js -o test/browser/di.js' && bash -c 'node_modules/.bin/mocha-phantomjs test/browser/browser.html' && rm test/browser/di.js",
-    "doc": "bash -c 'node_modules/.bin/jsdoc di.js -t node_modules/docdash/ -d doc'",
-    "build-dist": "bash -c 'node_modules/.bin/browserify di.js --standalone Di -o dist/di.js'",
-    "jshint": "bash -c 'node_modules/jshint/bin/jshint di.js test/di.js'",
+    "doc": "bash -c 'node_modules/.bin/jsdoc src/di.js -t node_modules/docdash/ -d doc'",
+    "build-dist": "bash -c 'mkdir -p dist' && bash -c 'node_modules/.bin/browserify src/di.js --standalone Di -o dist/di.js'",
+    "jshint": "bash -c 'node_modules/jshint/bin/jshint src/di.js test/di.js'",
     "build": "npm run build-dist && npm run jshint && npm run test && npm run doc",
     "prepublish": "npm run build"
   },
